requirements.txt
streamlit==1.32.0
langchain==0.1.0
langchain-community==0.0.10
pypdf2==3.0.1
pdfplumber==0.10.3
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.24.3
pandas==2.0.3
plotly==5.18.0
rouge-score==0.1.2
bert-score==0.3.13
chromadb==0.4.22
openai==1.6.1  # Opcional, si usas OpenAI